Support for calling lexers on individual matches
Changes the `:__this__` special name to `:root`, which retains the
same behaviour as before, but now extends to arbitrary states in the
current lexer.

Adds proper support for calling different lexers on a specific match
using `OtherLexer` as the token 'binding'.

Also adds more exact control for state and lexer changes with
`OtherLexer => :other_state` syntax.